+++ src/mappers/mapper7.rs
@@ -35,7 +35,7 @@ impl Mapper for Mapper7 {
 
     fn get_prg_byte(&mut self, address: u16) -> u8 {
         if address >= 0x8000 {
-            let bank = (self.register & 7) as usize;
+            let bank = self.register & 7;
             self.mapper_internal.get_prg_rom_byte(address, bank, _32KB)
         } else {
             0
